+++ pype/lib/__init__.py
@@ -89,6 +89,8 @@ from .applications import (
 from .profiles_filtering import filter_profiles
 
 from .plugin_tools import (
+    TaskNotSetError,
+    get_subset_name,
     filter_pyblish_plugins,
     source_hash,
     get_unique_layer_name,
@@ -181,6 +183,8 @@ __all__ = [
 
     "filter_profiles",
 
+    "TaskNotSetError",
+    "get_subset_name",
     "filter_pyblish_plugins",
     "source_hash",
     "get_unique_layer_name",
